add: handlers package add: handler for book requests fix: remove handling code from main file
